Fix bugs

+++ dev/sql/create_tables.sql
@@ -172,13 +172,13 @@ CREATE TABLE short_options (
 CREATE TABLE emission (
   id                        SERIAL PRIMARY KEY,
   currency_id               INT UNSIGNED NOT NULL,
-  initial_value             INT UNSIGNED NOT NULL,
-  current_value             INT UNSIGNED NOT NULL,
-  percent_rate              INT UNSIGNED NOT NULL,
-  `start`                   INT UNSIGNED NOT NULL,
-  start_percent             INT UNSIGNED NOT NULL,
-  end_percent               INT UNSIGNED NOT NULL,
-  delta_percent             INT UNSIGNED NOT NULL
+  initial_value             BIGINT UNSIGNED NOT NULL,
+  current_value             BIGINT UNSIGNED NOT NULL,
+  percent_rate              BIGINT UNSIGNED NOT NULL,
+  `start`                   BIGINT UNSIGNED NOT NULL,
+  start_percent             BIGINT UNSIGNED NOT NULL,
+  end_percent               BIGINT UNSIGNED NOT NULL,
+  delta_percent             BIGINT UNSIGNED NOT NULL
 ) DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;
